api/pyproject.toml
[project]
name = "sh-rag-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.15.0",
    "einops>=0.8.0",
    "lancedb>=0.17.0",
    "langchain-community>=0.3.14",
    "langchain-openai>=0.2.14",
    "langchain>=0.3.14",
    "mammoth>=1.8.0",
    "omegaconf>=2.3.0",
    "pypdf2>=3.0.1",
    "rich>=13.9.4",
    "streamlit-authenticator>=0.4.1",
    "streamlit>=1.41.1",
    "wheel>=0.45.1",
    "psutil>=6.1.1",
    "bentoml>=1.3.19",
    "asyncio>=3.4.3",
    "ninja>=1.11.1.2",
    "transformers>=4.47.0",
    "tantivy>=0.22.0",
    "langgraph>=0.2.63",
    "fastapi[standard]>=0.115.8",
    "pyjwt>=2.10.1",
    "bcrypt>=4.2.1",
    "sqlmodel>=0.0.22",
]

[project.optional-dependencies]
compile = ["flash-attn==2.7.2.post1"]

[tool.uv]
no-build-isolation-package = ["flash-attn"]

[[tool.uv.dependency-metadata]]
name = "flash-attn"
version = "2.7.2.post1"
requires-dist = ["torch", "einops", "setuptools"]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu124", marker = "platform_system != 'Darwin'"},
]
torchvision = [
    { index = "pytorch-cpu", marker = "platform_system == 'Darwin'" },
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

fixable = ["ALL"]


line-length = 88
indent-width = 4

# Assume Python 3.12
target-version = "py312"

[tool.ruff.lint]
# Add rules to enable
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
]

# Ignore specific rules
ignore = ["E501"]

[tool.ruff.format]
# Use double quotes for strings.
quote-style = "double"

# Indent with spaces, rather than tabs.
indent-style = "space"

# Respect magic trailing commas.
skip-magic-trailing-comma = false

# Automatically detect the appropriate line ending.
line-ending = "auto"

docstring-code-format = true

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.ruff.lint.isort]
# Group imports by type (standard library, third-party, first-party).
combine-as-imports = true
force-wrap-aliases = true
